Make user side menu a drawer navigator
