+++ SnowRide/App.js
@@ -1,98 +1,30 @@
-import { StatusBar } from "expo-status-bar";
-import { StyleSheet, Text, View } from "react-native";
+import { StyleSheet } from "react-native";
+import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { NavigationContainer } from "@react-navigation/native";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import React from "react";
+// import Navigator from "./Routes/userDrawer";
 
-import DriverPost from "./Screens/DriverPost";
-import PassengerPost from "./Screens/PassengerPost";
 import Welcome from "./Screens/Welcome";
 import User from "./Screens/User";
 import ChatBox from "./Screens/ChatList";
 
+import PostOverview from "./Routes/homeStack";
 import LoginScreen from "./Components/User/Login";
 import SignUpScreen from "./Components/User/SignUp";
 import UserPost from "./Components/User/UserPost";
 import UserProfile from "./Components/User/UserProfile";
 import PostForm from "./Components/Post/ManageEntry/PostForm";
-// import AddPost from "./Screens/AddPost";
 
-import {
-  SimpleLineIcons,
-  Ionicons,
-  MaterialCommunityIcons,
-} from "@expo/vector-icons";
 import { Colors } from "./Constants/colors";
 import PostDetail from "./Components/Post/PostDetail/PostDetail";
 import EditProfile from "./Screens/EditProfile";
-
-function PostOverview() {
-  const BottomTabs = createBottomTabNavigator();
-  return (
-    <BottomTabs.Navigator
-      screenOptions={({ navigation }) => ({
-        headerTintColor: Colors.tertiary100,
-        headerStyle: { backgroundColor: Colors.primary100 },
-        tabBarStyle: { backgroundColor: Colors.primary100 },
-        tabBarActiveTintColor: Colors.tertiary100,
-        tabBarInactiveTintColor: Colors.secondary200,
-        headerRight: ({ tintColor }) => (
-          <Ionicons
-            name="chatbox-ellipses-outline"
-            size={22}
-            color={tintColor}
-            onPress={() => {
-              navigation.navigate("Messages");
-            }}
-          />
-        ),
-        headerLeft: ({ tintColor }) => (
-          <SimpleLineIcons
-            name="user"
-            size={22}
-            color={tintColor}
-            onPress={() => {
-              navigation.navigate("User");
-            }}
-          />
-        ),
-      })}
-    >
-      <BottomTabs.Screen
-        name="Driver Posts"
-        component={DriverPost}
-        options={{
-          tabBarLabel: "DriverPosts",
-          tabBarIcon: ({ color, size }) => {
-            return <Ionicons name="car" size={size} color={color} />;
-          },
-        }}
-      />
-      <BottomTabs.Screen
-        name="Passenger Posts"
-        component={PassengerPost}
-        options={{
-          tabBarLabel: "PassengerPosts",
-          tabBarIcon: ({ color, size }) => {
-            return (
-              <MaterialCommunityIcons
-                name="car-seat"
-                size={size}
-                color={color}
-              />
-            );
-          },
-        }}
-      />
-    </BottomTabs.Navigator>
-  );
-}
+import RootDrawerNavigator from "./Routes/userDrawer";
 
 export default function App() {
   const Stack = createNativeStackNavigator();
   return (
-    <>
-      {/* <StatusBar style="auto" /> */}
+    <SafeAreaProvider>
       <NavigationContainer
         screenOptions={{
           headerStyle: { backgroundColor: Colors.secondary100 },
@@ -129,6 +61,8 @@ export default function App() {
           <Stack.Screen name="User" component={User} />
           <Stack.Screen name="UserPosts" component={UserPost} />
           <Stack.Screen name="UserProfile" component={UserProfile} />
+          <Stack.Screen name="UserDrawer" component={RootDrawerNavigator} />
+
 
           <Stack.Screen
             name="AddPost"
@@ -150,10 +84,9 @@ export default function App() {
             }}
           />
           <Stack.Screen name="Messages" component={ChatBox} />
-          {/* <Stack.Screen name="AddPost" component={AddPost} /> */}
         </Stack.Navigator>
       </NavigationContainer>
-    </>
+    </SafeAreaProvider>
   );
 }
