+++ SnowRide/Components/User/DrawerHeader.js
@@ -1,0 +1,10 @@
+import React from "react";
+import { StyleSheet, Text, View } from "react-native";
+
+export default function DrawerHeader() {
+  return (
+    <View style={styles.container}>
+      <Text>DrawerHeader</Text>
+    </View>
+  );
+}
