+++ SnowRide/babel.config.js
@@ -3,6 +3,7 @@ module.exports = function (api) {
   return {
     presets: ["babel-preset-expo"],
     plugins: [
+      'react-native-reanimated/plugin',
       [
         "module:react-native-dotenv",
         {
