+++ SnowRide/index.js
@@ -1,0 +1,9 @@
+import { AppRegistry, Platform } from 'react-native';
+import App from './App';
+
+AppRegistry.registerComponent('main', () => App);
+
+if (Platform.OS === 'web') {
+    const rootTag = document.getElementById('root') || document.getElementById('main');
+    AppRegistry.runApplication('main', { rootTag });
+}
