+++ SnowRide/package.json
@@ -18,7 +18,7 @@
     "firebase": "^9.18.0",
     "import": "^0.0.6",
     "react": "18.2.0",
-    "react-native": "0.71.4",
+    "react-native": "0.71.6",
     "react-native-async-storage": "^0.0.1",
     "react-native-datepicker": "^1.7.2",
     "react-native-dotenv": "^3.4.8",
@@ -29,6 +29,7 @@
     "react-native-safe-area-view": "^1.1.1",
     "react-native-screens": "^3.20.0",
     "react-navigation": "^4.4.4",
+    "react-navigation-drawer": "^2.7.2",
     "react-navigation-stack": "^2.10.4"
   },
   "devDependencies": {
